Improves MatchJSON error message
When one of the actual or expected cannot be parsed as JSON, it
would be quite nice to know quickly which one you, kind developer,
should be looking at in order to improve and fix your tests.

This commit mostly just adds the words "expected" and "actual"
around the error messages, which is hopefully more clear.