fix BuildIn on go1.9
os/exec in go1.9 prevents child processes from running with duplicate
environment variables. In the case of duplicate env vars, os/exec will pick the
last value. This broke gexec since it used to prepend the new value of GOPATH
which caused the old GOPATH to win.

I tried to append GOPATH but that doesn't seem to work on go versions <1.9. I
decided it is better to use a defer instead of branching on the version of go