Add methods to terminate/kill sessions and wait
Added new methods for sessions to kill resp. terminate them and wait for
that to actually happen:
- KillAndWait
- TerminateAndWait

Main incentive for those is to make it a bit easier to implement parts
of the gexec.Session interface -- to fake certain aspects of it, without the
need of that fake to know about gexec.Session.