Add BeElementOf matcher
The BeElementOf matcher checks if actual is contained within the passed in
elements. BeElementOf() always uses Equal() to perform the match.

When the passed in elements are comprised of a single one element that is
either a Map or Array or Slice, it behaves as the reverse form of the
ContainElement() with the Equal() matcher.

Otherwise, it provides syntactic sugar for:
Or(Equals(1), Equals(2), ...)