allow users to wrap the Gomega interface
gomega internally casts to *internal.Gomega which makes wrapping
it impossible.  To allow this, add an interface that if implemented
will be used to get the actual *internal.Gomega.  This allows users
to replace gomega.Default with their own wrapper as long as it
implements provides an 'Inner() Gomega' method.