Fix max number of samples in experiments on non-64-bit systems.
The use of `math.MaxInt64` as the maximum number of samples in the context
of general int variables can lead to inconsistent behavior on non-64-bit
systems.

This patch addresses this by using `math.MaxInt32` as the default sample
size in case an explicit one is not provided in the `SamplingConfig`.

Signed-off-by: Nashwan Azhari <nazhari@cloudbasesolutions.com>