fix(async_assertion): use correct error in errors.As
In AsyncAssertion.match, there is a confusion between
actualErr/matcherErr in one of the branches:

Before this commit, the code is:
```
if actualErr == nil
	if matcherErr == nil {
		// actualErr is nil, matcherErr is nil
		…
	} else {
		// actualErr is nil, matcherErr is non-nil
		var fgErr formattedGomegaError
		if errors.As(actualErr, &fgErr) {
			message += fgErr.FormattedGomegaError() + "\n"
		} else {
			message += renderError(fmt.Sprintf("The matcher passed to %s returned the following error:", assertion.asyncType), matcherErr)
		}
	}
} else {
	// actualErr is non-nil
	…
}
```

Calling `errors.As` on the nil `actualErr` looks like a mistake, this commit
changes it to `matcherErr` which is non-nil. This also matches the error
used in the second branch of the `if errors.As()` test.

This was reported by a static analysis tool.

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>